+++ demo/admin/src/products/future/ProductVariantForm.cometGen.ts
@@ -1,0 +1,21 @@
+import { future_FormConfig as FormConfig } from "@comet/cms-admin";
+import { GQLProductVariant } from "@src/graphql.generated";
+
+export const ProductVariantForm: FormConfig<GQLProductVariant> = {
+    type: "form",
+    gqlType: "ProductVariant",
+    fragmentName: "ProductVariantFormDetails",
+    fields: [
+        {
+            name: "name",
+            type: "text",
+            label: "Name",
+        },
+        {
+            name: "image",
+            type: "block",
+            label: "Image",
+            block: { name: "DamImageBlock", import: "@comet/cms-admin" },
+        },
+    ],
+};