+++ packages/admin/admin-babel-preset/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @comet/admin-babel-preset
 
+## 7.6.0
+
 ## 7.5.0
 
 ## 7.4.2